# Memoize the result of gsubbing @virtual_path