@pixeltrix 2018-03-20
@oniofchaos do you have any performance stats to back this change up? There's a lot of work going on to implement the cache and I'm unsure how much you're actually gaining.