@oniofchaos 2018-03-20
In my local testing, I found that this code path got executed twice for each call to `I18n.t` in my view. This PR reduces the `gsub` calls from once per call to once per namespace. 

Some raw benchmarking stats:
```
master_version
{:FREE=>-4568, :T_STRING=>2053, :T_MATCH=>2000, :T_IMEMO=>1000}
fast_version
{:FREE=>-1008, :T_STRING=>1003, :T_HASH=>1, :T_MATCH=>2, :T_IMEMO=>1}
Warming up --------------------------------------
      master_version    24.453k i/100ms
        fast_version    97.766k i/100ms
Calculating -------------------------------------
      master_version    283.462k (± 6.3%) i/s -      1.418M in   5.022541s
        fast_version      1.413M (±10.6%) i/s -      7.039M in   5.048207s

Comparison:
        fast_version:  1412516.5 i/s
      master_version:   283462.1 i/s - 4.98x  slower
```

Code used to generate
```ruby
# frozen_string_literal: true
begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update
                your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"

  gem "benchmark-ips"
  gem "rails"
end

def allocate_count
  GC.disable
  before = ObjectSpace.count_objects
  yield
  after = ObjectSpace.count_objects
  after.each { |k,v| after[k] = v - before[k] }
  after[:T_HASH] -= 1 # probe effect - we created the before hash.
  GC.enable
  result = after.reject { |k,v| v == 0 }
  GC.start
  result
end

def master_version(virtual_path, key)
  "#{virtual_path.gsub(%r{/_?}, ".")}#{key}"
end

def fast_version(virtual_path, key)
  @_cache ||= {}
  @_cache[virtual_path] ||= "#{virtual_path.gsub(%r{/_?}, ".")}#{key}"
  "#{@_cache[virtual_path]}#{key}"
end

def test
  puts "master_version"
  puts allocate_count { 1000.times { master_version('test/namespace/index', 'cat') } }
  puts "fast_version"
  puts allocate_count { 1000.times { fast_version('test/namespace/index', 'cat') } }

  Benchmark.ips do |x|
    x.report("master_version")  { master_version('test/namespace/index', 'cat') }
    x.report("fast_version")    { fast_version('test/namespace/index', 'cat') }
    x.compare!
  end
end

test
```