@pixeltrix 2018-03-20
Numbers seem good then - my only concern is that `@_scope_cache` is a fairly generic name and might be clobbered (or clobber) an instance variable from an application's helpers. Maybe use `@_i18n_scope` instead?