@oniofchaos 2018-03-20
Yeah good point about the generic name. What do you think about `@_ scope_key_by_partial_cache`? Longer name but even less likely for clobbering.